+++ BryanChi_FX Devices/types.lua
@@ -1,0 +1,5 @@
+---@meta
+
+
+---@alias Style "Pro C 2"|"Pro C Thresh"|"Custom Image"|"Invisible"|"FX Layering"|'up-down arrow'
+---@alias Position "Top"|"Free"|"Bottom"|"Within"|"Left"|"None"|"Right"
